Movement and Camera Start
Began working on movement and camera controls for the player.
Currently not fully functioning as the player will move based on the
rotation of the camera.
The camera needs to be alligned in a way such that it does not rotate
around its Z-Axis.
Will need to adjust how the camera's rotation will impact the movement
of the player.